+++ test/components/dropdown/dropdown.e2e-spec.js
@@ -808,3 +808,29 @@ describe('Dropdown "No Search" stay-open behavior', () => {
     expect(await element.all(by.css('div.dropdown span')).first().getText()).toBe('');
   });
 });
+
+describe('Dropdown blank option tests', () => {
+  it('Highlights the blank option when the list is opened', async () => {
+    await utils.setPage('/components/dropdown/test-blank-initially.html?layout=nofrills');
+    await utils.checkForErrors();
+
+    const dropdownEl = await element(by.css('div.dropdown'));
+    await dropdownEl.click();
+    await browser.driver
+      .wait(protractor.ExpectedConditions.presenceOf(element(by.id('dropdown-list'))), config.waitsFor);
+
+    expect(await element(by.css('.dropdown-option.is-selected')).getAttribute('data-val')).toBe('blank');
+  });
+
+  it('Highlights the blank option when the list is opened and virtual scrolling is enabled', async () => {
+    await utils.setPage('/components/dropdown/test-virtual-scroll-blank-option.html?layout=nofrills');
+    await utils.checkForErrors();
+
+    const dropdownEl = await element(by.css('div.dropdown'));
+    await dropdownEl.click();
+    await browser.driver
+      .wait(protractor.ExpectedConditions.presenceOf(element(by.id('dropdown-list'))), config.waitsFor);
+
+    expect(await element(by.css('.dropdown-option.is-selected')).getAttribute('data-val')).toBe('blank');
+  });
+});